feat: divided issuer/verifier
